requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asyncio==3.4.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
importlib-metadata==4.10.0
pyee==8.2.2
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
tornado==6.1
tqdm==4.62.3
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
websockets==9.1
zipp==3.6.0
